% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareBamFiles.R
\name{bamFileListToNanoblotData}
\alias{bamFileListToNanoblotData}
\title{Converting BamFileList object to Nanoblot dataframe suitable for Nanoblot analysis}
\usage{
bamFileListToNanoblotData(BamFileList)
}
\arguments{
\item{BamFileList}{A BamFileList object from the Rsamtools class}
}
\description{
Extracts relevant fields from BamFile within BamFileList and converts into a dataframe format

Note: All .bam files are loaded into RAM. This can result in issues if the
bam files are large. Make sure that bam files are subset already through subsetNanoblot

All paths and names of BamFileList object must be unique. Names of each element of BamFileList entry will be automatically
determined based on file path. User can override by using names(BamFileList) <- c()
}
\examples{
TestBamFileList <- Rsamtools::BamFileList(c("./temp/WT_sorted_merged_RPL18A_Exon1.bam", "./temp/RRP6_sorted_merged_RPL18A_Exon1.bam"))
names(TestBamFileList) <- c("WT","RRP6")
NanoblotDataWTRRP6 <- bamFileListToNanoblotData(TestBamFileList)


}
