% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeNanoblot.R
\name{calculateDESeqSizeFactors}
\alias{calculateDESeqSizeFactors}
\title{Calculate DESeq Size Factors using DESeq2}
\usage{
calculateDESeqSizeFactors(nanoblotData, unnormalizedFiles, annotationFile)
}
\arguments{
\item{nanoblotData}{This is a data frame that is generated from bamFilesToNanoblotData}

\item{unnormalizedFiles}{This is a vector that contains the string file locations of all the samples in nanoblotData}

\item{annotationFile}{This is a gtf file obtained from a database that contains the organism of interest's genomic annotations}
}
\description{
This function is required to normalize Nanoblots using the DeSeq normalization method.
It is not meant to be used by the end user.
It uses a GTF annotation file and
creates a count table using unnormalized files which is then passed to DeSeq2's sizeFactors() function
}
\examples{
TestDataNames <- c("WT_RPL18A", "RRP6_RPL18A")
TestDataFiles <- c("./temp/WT_RPL18A_Exon1.bam", "./temp/RRP6_RPL18A_Exon1.bam")
annotation <- "./user_input_files/Saccharomyces_cerevisiae.R64-1-1.107.gtf"
unnormalizedLocations <- c("./data/example/WT_sorted_merged.bam", "./data/example/RRP6_sorted_merged.bam")

BamsWTRRP6 <- scanBamFiles(TestDataNames, TestDataFiles)
NanoblotDataWTRRP6 <- bamFilesToNanoblotData(BamsWTRRP6)
calculateDESeqSizeFactors(NanoblotDataWTRRP6, unnormalizedLocations, annotation)

returns a vector with 2 elements
}
