% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeNanoblot.R
\name{calculateDuplicationFactors}
\alias{calculateDuplicationFactors}
\title{Calculate Duplication Factors for NanoBlot based off of size factors for plot type "blot"}
\usage{
calculateDuplicationFactors(
  nanoblotData,
  size_factors,
  DUPLICATION_CONSTANT = 2000
)
}
\arguments{
\item{nanoblotData}{This is a data frame that is generated from bamFilesToNanoblotData}

\item{size_factors}{This is a vector that contains the size factors for each sample in the nanoblotData}

\item{DUPLICATION_CONSTANT}{This is essentiall a value that determines how many "points" a standard Nanoblot should have.
The higher this number, the greater the opacity of the visualization. Default value is set to 2000}
}
\description{
This function is required to plot Nanoblots of type "blot" by creating duplication factors to standardize
visualization opacity. It is not meant to be used by the end user.
This function is called in the makeNanoblot function when the plot type of "blot" is specified
}
\examples{
TestDataNames <- c("WT_RPL18A", "RRP6_RPL18A")
TestDataFiles <- c("./temp/WT_RPL18A_Exon1.bam", "./temp/RRP6_RPL18A_Exon1.bam")
unnormalizedLocations <- c("./data/example/WT_sorted_merged.bam", "./data/example/RRP6_sorted_merged.bam")

BamsWTRRP6 <- scanBamFiles(TestDataNames, TestDataFiles)
NanoblotDataWTRRP6 <- bamFilesToNanoblotData(BamsWTRRP6)
size_factors <- calculateDESeqSizeFactors(NanoblotDataWTRRP6, unnormalizedLocations)
calculateDuplicationFactors(NanoblotDataWTRRP6, size_factors)

returns a vector with 2 elements of data type containing only whole numbers
}
