% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkIntegrity.R
\name{calculateIntegrity}
\alias{calculateIntegrity}
\title{Calculate Integrity}
\usage{
calculateIntegrity(GeneTargets, DataSet)
}
\arguments{
\item{GeneTargets}{The result of running bamFilesToNanoblotData}

\item{DataSet}{An optional title for the resulting plot.}
}
\description{
Calculate Integrity
}
\examples{
makeNanoblot()


}
