% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkForMultimapped.R
\name{checkForMultimapped}
\alias{checkForMultimapped}
\title{Check for Multimapping}
\usage{
checkForMultimapped(BamFileLocation, WarnPercent = 10)
}
\arguments{
\item{BamFileLocation}{The path to the BAM file to be
read in. The BAM files is imported to R via Rsamtools::scanBam.}

\item{WarnPercent}{A number between 0 and 100. If more then the specified
percent are reads in a BAM file are specified as secondary alignments then
a warning is issued.}
}
\description{
Check for Multimapping
}
\examples{
checkForMultimapped("Path/to/bam/file.bam")



}
