% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkIntegrity.R
\name{checkIntegrity}
\alias{checkIntegrity}
\title{Check Integrity of Bam files for Nanoblot}
\usage{
checkIntegrity(GeneTargets, BamFiles)
}
\arguments{
\item{GeneTargets}{A GRanges object with the target genomic regions.
These genomic regions should theoretically have no 5' transcripts ends.}

\item{BamFiles}{A BamFileList() from Rsamtools}
}
\description{
Checks the percent of intact reads for all gene targets and then outputs a scatter plot of integrity
along with the cumulative distribution plot.
}
\examples{
makeNanoblot()

YeastCDSGFF <- rtracklayer::import.gff("~/Downloads/saccharomyces_cerevisiae.20210411.CDS.no_mito.gff")
TestBamFileList <- Rsamtools::BamFileList(c("./../data/example/WT_sorted_merged.bam", "./../data/example/RRP6_sorted_merged.bam"))

counts <- calculateIntegrity(GeneTargets = YeastCDSGFF, BamFiles = TestBamFileList)
}
