% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeNanoblot.R
\name{duplicateNanoblotData}
\alias{duplicateNanoblotData}
\title{Duplicate NanoblotData for the purposes of "blot" type visualization}
\usage{
duplicateNanoblotData(nanoblotData, duplication_factors)
}
\arguments{
\item{nanoblotData}{This is a data frame that is generated from bamFilesToNanoblotData}

\item{duplication_factors}{This is a vector that contains the duplication factors for each sample in the nanoblotData and generated
from the function calculateDuplicationFactors}
}
\description{
This function is required to plot Nanoblots of type "blot" by duplicating existing Nanoblot data for
visualization opacity. It is not meant to be used by the end user.
This function is called in the makeNanoblot function when the plot type of "blot" is specified
}
\examples{
TestDataNames <- c("WT_RPL18A", "RRP6_RPL18A")
TestDataFiles <- c("./temp/WT_RPL18A_Exon1.bam", "./temp/RRP6_RPL18A_Exon1.bam")
unnormalizedLocations <- c("./data/example/WT_sorted_merged.bam", "./data/example/RRP6_sorted_merged.bam")

BamsWTRRP6 <- scanBamFiles(TestDataNames, TestDataFiles)
NanoblotDataWTRRP6 <- bamFilesToNanoblotData(BamsWTRRP6)
size_factors <- calculateDESeqSizeFactors(NanoblotDataWTRRP6, unnormalizedLocations)
duplication_factors <- calculateDuplicationFactors(NanoblotDataWTRRP6, size_factors)
duplicateNanoblotData(NanoblotDataWTRRP6, duplication_factors)

returns the original NanoblotDataWTRRP6 data frame with duplicated data for each sample
}
