% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{extractNanoblotData}
\alias{extractNanoblotData}
\title{Extract Nanoblot Data from Rsamtools::scanBam()}
\usage{
extractNanoblotData(SampleList, SampleNames, index)
}
\arguments{
\item{SampleList}{A list of the outputs of Rsamtools::scanBam() to extract the qname and qwidth from}
}
\description{
This function is required to run NanoBlot. It is not intended for use by an end user.
}
\examples{
extractNanoblotData(SampleList)

returns a dataframe with 2 columns
}
