% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{isUnique}
\alias{isUnique}
\title{Unique Vector Checker}
\usage{
isUnique(vector)
}
\arguments{
\item{vector}{Any vector to check uniqueness on.
Returns true if all elements are unique.}
}
\description{
This function is required to run NanoBlot. It is not intended
for use by an end user. This function takes a vector input and checks that each element
in that vector is Unique.
}
\examples{
testVector_1 <- c('alpha', 'beta', 'gamma')
testVector_2 <- c('alpha', 'beta', 'alpha')

isUnique(testVector_1)
  returns TRUE
isUnique(testVector_2)
  returns FALSE
}
