% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeNanoblot.R
\name{makeNanoblot}
\alias{makeNanoblot}
\title{Generate Nanoblot from NanoblotData}
\usage{
makeNanoblot(
  nanoblotData,
  plotInfo,
  blotType = "blot",
  plotTitle = "",
  size_factors = rep(1, length(levels(nanoblotData$SampleID)))
)
}
\arguments{
\item{nanoblotData}{The result of running bamFilesToNanoblotData}

\item{plotInfo}{A data frame.with 3 columns of equal lengths called SampleID, SampleLanes and, SampleColors}

\item{blotType}{The type of Nanoblot to produce. There are 3 acceptable options: 'blot', 'violin', and 'ridge'. Default 'blot'}

\item{plotTitle}{An optional title for the resulting plot. Default blank}
}
\description{
This function takes a NanoblotData object, a plotInfo dataframe, and
returns a ggplot object
}
\examples{

ExampleDataset <- bamFilesToNanoblotData()

ExamplePlotInfo <- data.frame(
  SampleID = c(WT,Test),
  SampleLanes = c(1,2),
  SampleColors = c('Red','Blue')
    )

makeNanoblot(nanoblotData = ExampleDataset,
plotInfo = ExamplePlotInfo,
plotTitle = "Test plot",
)

}
