% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareBamFiles.R
\name{scanBamFiles}
\alias{scanBamFiles}
\title{Reading in bam files with scanBam}
\usage{
scanBamFiles(SampleID, BamFileLocations)
}
\arguments{
\item{SampleID}{A vector of unique sample names.}

\item{BamFileLocations}{A vector containing the paths to the bam files to be
read in. .bam files are imported to R via Rsamtools::scanBam.}
}
\description{
A wrap around function which runs Rsamtools::scanBam on a series
of .bam files.

Note: All .bam files are loaded into RAM. This can result in issues if the
bam files are large.
}
\examples{
sampleNames <- c('WT', 'test')
sampleLocations <- c('/path/to/WT.bam', '/path/to/test.bam')

SampleBamFiles <- scanBamFiles(SampleID=sampleNames, BamFileLocations=sampleLocations)

SampleNanoblotData <- bamFilesToNanoblotData(SampleBamFiles)

}
