% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetNanoblot.R
\name{subsetNanoblot}
\alias{subsetNanoblot}
\title{Subsetting Bam files for Nanoblot generation}
\usage{
subsetNanoblot(
  BamFileList,
  probesFile,
  targetProbes,
  targetAntiProbes = NULL,
  viewingWindow = NULL,
  cDNA = FALSE,
  RTPCR = FALSE,
  RACE = FALSE,
  tempFilePath = "./temp"
)
}
\arguments{
\item{BamFileList}{Takes a BamFileList object type from the RSamtools class. If user wants to specify naming of the samples, make sure
to use the names(BamFileList) accessor function and assign names. Names and file paths must be unique.}

\item{probesFile}{Takes a .bed traditional BED file format with 6 columns. chrom, chromStart, chromEnd, name, score, and strand.
File must be tab delimited as well.}

\item{targetProbes}{Vector of target probes thare are located in the probesFile argument}

\item{targetAntiProbes}{Vector of target anti probes that are located in the probesFile argument. This is default to NULL}

\item{viewingWindow}{Viewing window for RT-PCR and RACE modes. Viewing window must befound in probesFile}

\item{cDNA}{Boolean paramter to determines whether to treat reads as cDNA or not. Default to FALSE}

\item{RTPCR}{Boolean parameter to determine whether to use RT-PCR mode for viewing window. Default to FALSE}

\item{RACE}{Boolean parameter to determine whether to use RACE mode. RT-PCR must be TRUE for RACE to work. Default to FALSE}

\item{tempFilePath}{File path to specify where temp bam subset files will be stored. Default is ./temp folder in the
current working directory. If the directory does not exist, this function will create it.}
}
\description{
Essentially calls bedtools intersect and samtools ampliconclip with various different parameters
}
\examples{

TestBamFileList <- Rsamtools::BamFileList(c("./data/example/WT_sorted_merged.bam","./data/example/RRP6_sorted_merged.bam"))
subsetNanoblot(TestBamFileList, "./user_input_files/probes.bed", c("RPL18A_Exon1"))

}
